finally, working multi-diff functionality :D
quick explanation of how it works - the multi_diff method takes 2 args: base_text (the text of the element in the elements_data table) and an array edits (contains the "data" column in elements_data_approve table for all requested approvals). first, all the changes in edits are merged (this is done by diffing them amongst each other). the merged result is then diffed against the base_text.

note that this method works for both single- and multi-diffs.

in the words of Ali G - booyakasha! :P